Add orientation view

+++ AVCam/MadgwickAHRS.h
@@ -1,13 +1,13 @@
 class MadgwickAHRS{
 	private:
-		float beta;				// algorithm gain
-		float q0, q1, q2, q3;	// quaternion of sensor frame relative to auxiliary frame
-//		float M_PI;
-		float lastUpdate;
+		double beta;				// algorithm gain
+		double q0, q1, q2, q3;	// quaternion of sensor frame relative to auxiliary frame
+//		double M_PI;
+		double lastUpdate;
 
 	public:
 		MadgwickAHRS();
-		void MadgwickAHRSupdateIMU(float gx, float gy, float gz, float ax, float ay, float az, float nanoTime);
-		void Quaternion2YPR(float * angles);
-		float invSqrt(float x);
+		void MadgwickAHRSupdateIMU(double gx, double gy, double gz, double ax, double ay, double az, double nanoTime);
+		void Quaternion2YPR(double * angles);
+		double invSqrt(double x);
 };
